requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
pydub==0.25.1
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
requests==2.28.2
selenium==4.8.0
selenium-recaptcha-solver==1.6.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.9.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
